record a test's module when storing group stats
The pytest plugin (`pytest_plugin.py`) has a global variable,
`test_group_stats`, which collects the group stats from each test ran.
It is populated in `pytest_generate_tests` and consumed in
`pytest_terminal_summary` to create the `results.json` file that
Gradescope needs. Previously, the key into `test_group_stats` was the
name of the test and graded region formatted into a string. However,
this caused an issue when two Python files containing tests for pytest
to run had a test by the same name. This meant that the later test
trampled on the group stats of the earlier test, even though the two
tests were completely separate and independent of each other. To remedy
this problem, I added the test file's name into the formatted string
that's used as a key into `pytest_terminal_summary`.